+++ src/decorator/Body.ts
@@ -1,15 +1,15 @@
 import { BodyOptions } from '../decorator-options/BodyOptions';
 import { getMetadataArgsStorage } from '../index';
-import { Newable } from '../types/Types';
+import { GenericObject } from '../types/Types';
 
-type BodyFunction = (object: Newable, methodName: string, index: number) => void;
+type BodyFunction = (object: GenericObject, methodName: string, index: number) => void;
 
 /**
  * Allows to inject a request body value to the controller action parameter.
  * Must be applied on a controller action parameter.
  */
 export function Body(options?: BodyOptions): BodyFunction {
-  return function (object: Newable, methodName: string, index: number) {
+  return function (object: GenericObject, methodName: string, index: number) {
     getMetadataArgsStorage().params.push({
       type: 'body',
       object: object,