+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### next
 
 - multer updated to `1.4.5-lts.1`
+- Upgraded to TypeScript 5.0.3 - @Body decorator type fixed
 
 ### 1.0.6